vinyl: rework vy_index reference counting
A vy_index might be dropped while there are still tasks pending for it.
To avoid use-after-free we currently use a rather weird technique: the
drop operation decrements vy_index->refs, but doesn't schedule index
destruction - instead the dead index is added to the "shutdown" list and
the scheduler itself keeps polling it queueing indexes whose ref counter
dropped to zero for destruction. Let's simplify it: instead of polling,
let's queue an index drop task as soon as index->refs drops to 0. This
way we don't need the "shutdown" list at all.

Note, that this patch removes the code destroying dead indexes from the
scheduler shutdown procedure. However, we don't release resources (open
files and memory) associated with indexes which are still alive, so
handling only dead indexes is a half-measure and keeping it doesn't look
reasonable. We might want to release all indexes on shutdown one day,
but that's a different story.

Also note there's no point in protecting the index ref counter with a
mutex as it is only modified from the tx thread hence ref_lock is
removed too.