vinyl: move vy_range_complete to worker
When we create a new range file, we first write a new run to a temporary
file from a worker thread and only then give it a proper name from the
tx thread. There were two reasons for doing this:

 1. On compaction, we used to add new ranges to the tree only after
    writing range files, so that w/o renaming a new range file could be
    garbage collected before compaction completion.
 2. Normally, renaming rules out incomplete range files.

Reason #1 isn't actual any longer since compaction was reworked so that
new ranges are added to the tree before creating any range files. This
prevents new range files from being garbage collected.

Reason #2 doesn't sound solid either, because we can append data to a
range (on dump) and therefore get a corrupted range as no renaming is
involved in such operation.

That said, we could get rid of renaming altogether. Nevertheless, let's
leave it for now, but just move renaming from the tx thread to a worker.
Rationale: renaming involves IO and hence had better be done from a
worker thread.